pyproject.toml
[project]
name = "proxy-mistral"
version = "0.1.0"
description = "Meeting Proxy Agent with Mistral AI"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "pipecat-ai[mistral,elevenlabs]>=0.0.100",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "websockets>=11.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "click>=8.0",
    "mistralai[realtime]>=1.0",
    "sqlite-utils>=3.0",
    "structlog>=23.0",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "soundfile>=0.12",
    "numpy>=1.24",
    "sounddevice>=0.4",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-auth-httplib2>=0.1",
    "python-jose>=3.3",
    "passlib>=1.7",
    "bcrypt>=4.0",
    "email-validator>=2.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.5",
    "itsdangerous>=2.0",
    "sentence-transformers>=2.2",
    "fugashi>=1.2",
    "unicodedata2>=15.0",
    "apscheduler>=3.10",
]

[project.scripts]
proxy-mistral = "src.main:cli"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "ruff>=0.1",
    "pip-audit>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
select = ["E", "F", "I"]

[tool.black]
line-length = 120
